refactored to cards
